# api/openapi.yaml
openapi: 3.1.0
info:
  title: Vermouth gateway
  version: 0.1.0
  description: |
    The single description of the gateway surface (STK-10). The gateway's Go
    request and response types are generated from this document with
    oapi-codegen, and the browser's types with openapi-typescript. A hand
    written client type for a gateway endpoint is not allowed.

    Only the gateway is described here. Services are reached by the gateway
    alone (INV-1), so their endpoints are not a public surface.

    The four auth endpoints are unauthenticated because they are what produces
    a token. Everything else under /api/ needs one. The teaching endpoints are
    the first real product thread through the gateway.
servers:
  - url: http://localhost:8080
    description: local development
paths:
...
  /api/classes:
    post:
      operationId: createClass
      summary: Create a class with its first session
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassRequest"
      responses:
        "200":
          description: The original class and first session for a safe retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateClassResponse"
        "201":
          description: The class and first session were created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateClassResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/classes/{class_id}/rates/{effective_date}:
    put:
      operationId: putClassRate
      summary: Create or correct one dated class rate
      parameters:
        - $ref: "#/components/parameters/ClassId"
        - name: effective_date
          in: path
          required: true
          schema:
            type: string
            format: date
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutClassRateRequest"
      responses:
        "200":
          description: The committed dated rate or its original retry response
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutClassRateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/billing-periods/default:
    get:
      operationId: getBillingPeriodDefault
      summary: Read the server derived most recent completed month
      responses:
        "200":
          description: Verified timezone and default completed period
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingPeriodDefault"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /api/billing-periods/{year}/{month}:
    get:
      operationId: getBillingPeriod
      summary: Read an unissued month or its immutable issued run
      parameters:
        - $ref: "#/components/parameters/BillingYear"
        - $ref: "#/components/parameters/BillingMonth"
      responses:
        "200":
          description: The current authoritative period state
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingPeriodState"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /api/billing-periods/{year}/{month}/preview:
    post:
      operationId: previewBillingPeriod
      summary: Calculate one completed month after proving projection readiness
      parameters:
        - $ref: "#/components/parameters/BillingYear"
        - $ref: "#/components/parameters/BillingMonth"
      responses:
        "200":
          description: Ready, blocked, empty, or already issued month
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingPreview"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "503":
          $ref: "#/components/responses/ProjectionUnavailable"
  /api/billing-periods/{year}/{month}/issue:
    post:
      operationId: issueBillingPeriod
      summary: Issue one immutable invoice per billable student
      parameters:
        - $ref: "#/components/parameters/BillingYear"
        - $ref: "#/components/parameters/BillingMonth"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueBillingRequest"
      responses:
        "200":
          description: The existing run returned for a retry
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingRun"
        "201":
          description: The new run and every immutable invoice
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingRun"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ProjectionUnavailable"
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: One browser generated UUID, retained until this create step succeeds.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    ClassId:
      name: class_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    BillingYear:
      name: year
      in: path
      required: true
      schema:
        type: integer
        minimum: 2000
    BillingMonth:
      name: month
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 12
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: An Ed25519 signed access token, verified locally by each service (INV-14).
  responses:
    BadRequest:
      description: The request was not acceptable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthenticated:
      description: No valid token was supplied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: No owned resource matched
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The requested state conflicts with committed state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnprocessableEntity:
      description: The request shape is valid but one or more profile fields are invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ProjectionUnavailable:
      description: The projection cannot yet prove the requested published cut
      headers:
        Retry-After:
          required: false
          schema:
            type: integer
            minimum: 1
        Cache-Control:
          required: true
          schema:
            type: string
            const: no-store
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      description: The one error shape at the gateway boundary (spec 0001).
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              examples: ["invalid_input"]
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
...
    ClassColor:
      type: string
      enum: [red, rose, orange, green, blue, yellow, violet]
    LocalTime:
      type: string
      pattern: "^(?:[01][0-9]|2[0-3]):[0-5][0-9]$"
      examples: ["17:30"]
    FirstSessionInput:
      type: object
      required: [local_date, start_time, end_time]
      properties:
        local_date:
          type: string
          format: date
        start_time:
          $ref: "#/components/schemas/LocalTime"
        end_time:
          $ref: "#/components/schemas/LocalTime"
    WeeklyScheduleSlotInput:
      type: object
      required: [weekday, start_time, end_time]
      properties:
        weekday:
          type: integer
          minimum: 1
          maximum: 7
        start_time:
          $ref: "#/components/schemas/LocalTime"
        end_time:
          $ref: "#/components/schemas/LocalTime"
    WeeklyScheduleInput:
      type: object
      required: [valid_from, valid_through, slots]
      properties:
        valid_from:
          type: string
          format: date
        valid_through:
          type: string
          format: date
        slots:
          type: array
          minItems: 1
          maxItems: 7
          items:
            $ref: "#/components/schemas/WeeklyScheduleSlotInput"
    CreateClassRequest:
      type: object
      description: Exactly one of first_session or schedule is required.
      required: [name, rate_amount]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        color:
          oneOf:
            - $ref: "#/components/schemas/ClassColor"
            - type: "null"
        rate_amount:
          type: integer
          format: int64
          minimum: 0
          maximum: 1000000000
        first_session:
          $ref: "#/components/schemas/FirstSessionInput"
        schedule:
          $ref: "#/components/schemas/WeeklyScheduleInput"
    Class:
      type: object
      required:
        [class_id, name, color, rate_amount, currency, rate_effective_from, schedule_revision, rate_revision]
      properties:
        class_id:
          type: string
          format: uuid
        name:
          type: string
        color:
          $ref: "#/components/schemas/ClassColor"
        rate_amount:
          type: integer
          format: int64
        currency:
          type: string
          const: VND
        rate_effective_from:
          type: string
          format: date
        schedule_revision:
          type: integer
          format: int64
          minimum: 0
        rate_revision:
          type: integer
          format: int64
          minimum: 0
    TeachingSession:
      type: object
      required: [session_id, class_id, starts_at, ends_at, local_date]
      properties:
        session_id:
          type: string
          format: uuid
        class_id:
          type: string
          format: uuid
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        local_date:
          type: string
          format: date
    CreateClassResponse:
      type: object
      required: [class, first_session, rule, candidate_count, created_count, adopted_count]
      properties:
        class:
          $ref: "#/components/schemas/Class"
        first_session:
          oneOf:
            - $ref: "#/components/schemas/TeachingSession"
            - type: "null"
        rule:
          oneOf:
            - $ref: "#/components/schemas/ScheduleRuleSummary"
            - type: "null"
        candidate_count:
          type: integer
          minimum: 0
        created_count:
          type: integer
          minimum: 0
        adopted_count:
          type: integer
          minimum: 0
...
    ScheduleRuleState:
      type: string
      enum: [planned, active, completed, replaced, ended, retired]
    ScheduleRuleSummary:
      type: object
      required:
        [schedule_rule_id, class_id, revision, valid_from, valid_through, time_zone, state, slots, replaced_at, ended_at, retired_at]
      properties:
        schedule_rule_id:
          type: string
          format: uuid
        class_id:
          type: string
          format: uuid
        revision:
          type: integer
          format: int64
          minimum: 1
        valid_from:
          type: string
          format: date
        valid_through:
          type: string
          format: date
        time_zone:
          type: string
        state:
          $ref: "#/components/schemas/ScheduleRuleState"
        slots:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyScheduleSlotInput"
        replaced_at:
          type: [string, "null"]
          format: date-time
        ended_at:
          type: [string, "null"]
          format: date-time
        retired_at:
          type: [string, "null"]
          format: date-time
...
    RateCurrent:
      type: object
      additionalProperties: false
      required: [rate_amount, currency, effective_from, rate_revision]
      properties:
        rate_amount:
          type: integer
          format: int64
          minimum: 0
          maximum: 1000000000
        currency:
          type: string
          const: VND
        effective_from:
          type: string
          format: date
        rate_revision:
          type: integer
          format: int64
          minimum: 0
    RateAllowedRange:
      type: object
      additionalProperties: false
      required: [from, through]
      properties:
        from:
          type: string
          format: date
        through:
          type: string
          format: date
...
    PutClassRateRequest:
      type: object
      additionalProperties: false
      required: [rate_amount]
      properties:
        rate_amount:
          type: integer
          format: int64
          minimum: 0
          maximum: 1000000000
    PutClassRateResponse:
      type: object
      additionalProperties: false
      required:
        [class_id, effective_from, rate_amount, currency, rate_revision, current, allowed_range, issued_invoices_unchanged, history_state]
      properties:
        class_id:
          type: string
          format: uuid
        effective_from:
          type: string
          format: date
        rate_amount:
          type: integer
          format: int64
        currency:
          type: string
          const: VND
        rate_revision:
          type: integer
          format: int64
          minimum: 1
        current:
          $ref: "#/components/schemas/RateCurrent"
        allowed_range:
          $ref: "#/components/schemas/RateAllowedRange"
        issued_invoices_unchanged:
          type: boolean
          const: true
        history_state:
          type: string
          const: pending
    BillingPeriod:
      type: object
      additionalProperties: false
      required: [year, month]
      properties:
        year:
          type: integer
          minimum: 2000
        month:
          type: integer
          minimum: 1
          maximum: 12
    BillingPeriodDefault:
      type: object
      additionalProperties: false
      required: [server_date, timezone, year, month, minimum_year]
      properties:
        server_date:
          type: string
          format: date
        timezone:
          type: string
        year:
          type: integer
          minimum: 2000
        month:
          type: integer
          minimum: 1
          maximum: 12
        minimum_year:
          type: integer
          const: 2000
    BillingRecoveryDestination:
      type: object
      additionalProperties: false
      required: [route, date, session_id, class_id, rate_date]
      properties:
        route:
          type: string
        date:
          type: [string, "null"]
          format: date
        session_id:
          type: [string, "null"]
          format: uuid
        class_id:
          type: [string, "null"]
          format: uuid
        rate_date:
          type: [string, "null"]
          format: date
    BillingBlocker:
      type: object
      additionalProperties: false
      required: [code, student_id, session_id, class_id, local_date, field, destination]
      properties:
        code:
          type: string
          enum: [profile_incomplete, attendance_incomplete, rate_missing]
        student_id:
          type: [string, "null"]
          format: uuid
        session_id:
          type: [string, "null"]
          format: uuid
        class_id:
          type: [string, "null"]
          format: uuid
        local_date:
          type: [string, "null"]
          format: date
        field:
          type: [string, "null"]
        destination:
          $ref: "#/components/schemas/BillingRecoveryDestination"
    BillingLine:
      type: object
      additionalProperties: false
      required: [session_id, class_id, class_name, local_date, rate_amount, amount, currency]
      properties:
        session_id:
          type: string
          format: uuid
        class_id:
          type: [string, "null"]
          format: uuid
        class_name:
          type: string
        local_date:
          type: string
          format: date
        rate_amount:
          type: integer
          format: int64
        amount:
          type: integer
          format: int64
        currency:
          type: string
          const: VND
    BillingStudentTotal:
      type: object
      additionalProperties: false
      required: [student_id, student_name, lines, total_amount, currency]
      properties:
        student_id:
          type: string
          format: uuid
        student_name:
          type: string
        lines:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/BillingLine"
        total_amount:
          type: integer
          format: int64
        currency:
          type: string
          const: VND
    IssuedInvoice:
      type: object
      additionalProperties: false
      required:
        [invoice_id, student_id, student_name, invoice_number, total_amount, currency, issued_at, lines]
      properties:
        invoice_id:
          type: string
          format: uuid
        student_id:
          type: string
          format: uuid
        student_name:
          type: string
        invoice_number:
          type: string
        total_amount:
          type: integer
          format: int64
        currency:
          type: string
          const: VND
        issued_at:
          type: string
          format: date-time
        lines:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/BillingLine"
    BillingRun:
      type: object
      additionalProperties: false
      required: [billing_run_id, period, generation, created_at, invoices, grand_total, currency]
      properties:
        billing_run_id:
          type: string
          format: uuid
        period:
          $ref: "#/components/schemas/BillingPeriod"
        generation:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/IssuedInvoice"
        grand_total:
          type: integer
          format: int64
        currency:
          type: string
          const: VND
    BillingPeriodState:
      type: object
      additionalProperties: false
      required: [status, period, run]
      properties:
        status:
          type: string
          enum: [unissued, already_issued]
        period:
          $ref: "#/components/schemas/BillingPeriod"
        run:
          oneOf:
            - $ref: "#/components/schemas/BillingRun"
            - type: "null"
    BillingPreview:
      type: object
      additionalProperties: false
      required: [status, period, students, blockers, grand_total, currency, preview_fingerprint, run]
      properties:
        status:
          type: string
          enum: [ready, blocked, empty, already_issued]
        period:
          $ref: "#/components/schemas/BillingPeriod"
        students:
          type: array
          items:
            $ref: "#/components/schemas/BillingStudentTotal"
        blockers:
          type: array
          items:
            $ref: "#/components/schemas/BillingBlocker"
        grand_total:
          type: integer
          format: int64
        currency:
          type: string
          const: VND
        preview_fingerprint:
          type: [string, "null"]
        run:
          oneOf:
            - $ref: "#/components/schemas/BillingRun"
            - type: "null"
    IssueBillingRequest:
      type: object
      additionalProperties: false
      required: [preview_fingerprint]
      properties:
        preview_fingerprint:
          type: string
          minLength: 1
...
security:
  - bearerAuth: []